Add streaming page again
This reverts to the streaming page that was removed in da53a475d8, but
updates it with the information for SotM 2025. It also adds some links
to other pages and VLC, a video player that can be used to play the
stream. In addition to that, there are .strm files for Kodi now, which
are also linked.